requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
aws-lambda-powertools==3.2.0
black==24.10.0
boto3==1.35.49
botocore==1.35.49
certifi==2024.8.30
cfgv==3.4.0
click==8.1.7
distlib==0.3.9
dnspython==2.7.0
dyntastic==0.15.0
email_validator==2.2.0
fastapi==0.115.3
filelock==3.16.1
flake8==7.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
identify==2.6.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
jmespath==1.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
pytest==8.3.3
pytest-pycharm==0.7.0
python-dateutil==2.9.0.post0
python-multipart==0.0.16
pytz==2024.2
PyYAML==6.0.2
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
starlette==0.41.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.27.0
aws-cdk-lib==2.164.1
constructs>=10.0.0,<11.0.0

